refactor (post): move form image to separate component
